+++ docs-util/oas-output/operations/admin/post_admin_campaigns_[id].ts
@@ -33,8 +33,6 @@
  * tags:
  *   - Campaigns
  * responses:
- *   "200":
- *     description: OK
  *   "400":
  *     $ref: "#/components/responses/400_error"
  *   "401":