+++ docs-util/oas-output/operations/admin/post_admin_customer-groups_[id]_customers_batch.ts
@@ -37,8 +37,6 @@
  * tags:
  *   - Customer Groups
  * responses:
- *   "200":
- *     description: OK
  *   "400":
  *     $ref: "#/components/responses/400_error"
  *   "401":