+++ packages/event-bus-redis/src/services/event-bus-redis.ts
@@ -20,8 +20,10 @@ export default class RedisEventBusService extends AbstractEventBusModuleService 
   protected readonly moduleOptions_: EventBusRedisModuleOptions
   // eslint-disable-next-line max-len
   protected readonly moduleDeclaration_: InternalModuleDeclaration
+  protected readonly eventBusRedisConnection_: Redis
 
   protected queue_: Queue
+  protected bullWorker_: Worker
 
   constructor(
     { logger, eventBusRedisConnection }: InjectedDependencies,
@@ -31,6 +33,8 @@ export default class RedisEventBusService extends AbstractEventBusModuleService 
     // @ts-ignore
     // eslint-disable-next-line prefer-rest-params
     super(...arguments)
+
+    this.eventBusRedisConnection_ = eventBusRedisConnection
 
     this.moduleOptions_ = moduleOptions
     this.logger_ = logger
@@ -44,12 +48,24 @@ export default class RedisEventBusService extends AbstractEventBusModuleService 
     // Register our worker to handle emit calls
     const shouldStartWorker = moduleDeclaration.worker_mode !== "server"
     if (shouldStartWorker) {
-      new Worker(moduleOptions.queueName ?? "events-queue", this.worker_, {
-        prefix: `${this.constructor.name}`,
-        ...(moduleOptions.workerOptions ?? {}),
-        connection: eventBusRedisConnection,
-      })
+      this.bullWorker_ = new Worker(
+        moduleOptions.queueName ?? "events-queue",
+        this.worker_,
+        {
+          prefix: `${this.constructor.name}`,
+          ...(moduleOptions.workerOptions ?? {}),
+          connection: eventBusRedisConnection,
+        }
+      )
     }
+  }
+
+  __hooks = {
+    onApplicationShutdown: async () => {
+      await this.bullWorker_?.close(true)
+      await this.queue_.close()
+      this.eventBusRedisConnection_.disconnect()
+    },
   }
 
   /**