+++ packages/medusa-test-utils/src/medusa-test-runner.ts
@@ -54,6 +54,7 @@ const dbTestUtilFactory = (): any => ({
   shutdown: async function (dbName: string) {
     await this.db_?.destroy()
     await this.pgConnection_?.context?.destroy()
+    await this.pgConnection_?.destroy()
 
     return await dropDatabase(
       { databaseName: dbName, errorIfNonExist: false },
@@ -104,39 +105,19 @@ export function medusaIntegrationTestRunner({
     debug,
   }
 
-  // Intercept call to this utils to apply the unique client url for the current suite
-  const originalCreatePgConnection =
-    require("@medusajs/utils/dist/modules-sdk/create-pg-connection").createPgConnection
-  require("@medusajs/utils/dist/modules-sdk/create-pg-connection").createPgConnection =
-    (options: any) => {
-      return originalCreatePgConnection({
-        ...options,
-        clientUrl: dbConfig.clientUrl,
-      })
-    }
-
-  const originalDatabaseLoader =
-    require("@medusajs/medusa/dist/loaders/database").default
-  require("@medusajs/medusa/dist/loaders/database").default = (
-    options: any
+  const originalConfigLoader =
+    require("@medusajs/medusa/dist/loaders/config").default
+  require("@medusajs/medusa/dist/loaders/config").default = (
+    rootDirectory: string
   ) => {
-    options.configModule.projectConfig.database_url
-    return originalDatabaseLoader({
-      ...options,
-      configModule: {
-        ...options.configModule,
-        projectConfig: {
-          database_logging: debug, // Will be used for the debug flag of the database options
-          ...options.configModule.projectConfig,
-          database_url: dbConfig.clientUrl,
-        },
-      },
-    })
+    const config = originalConfigLoader(rootDirectory)
+    config.projectConfig.database_url = dbConfig.clientUrl
+    return config
   }
 
   const cwd = process.cwd()
 
-  let shutdown: () => Promise<void>
+  let shutdown = async () => void 0
   let dbUtils = dbTestUtilFactory()
   let container: ContainerLike
   let apiUtils: any
@@ -162,6 +143,8 @@ export function medusaIntegrationTestRunner({
     getContainer: () => container,
   } as MedusaSuiteOptions
 
+  let isFirstTime = true
+
   const beforeAll_ = async () => {
     await dbUtils.create(dbName)
     const { dbDataSource, pgConnection } = await initDb({
@@ -176,7 +159,7 @@ export function medusaIntegrationTestRunner({
     dbUtils.pgConnection_ = pgConnection
 
     const {
-      shutdown: shutdown_,
+      shutdown: serverShutdown,
       container: container_,
       port,
     } = await startBootstrapApp({
@@ -184,13 +167,26 @@ export function medusaIntegrationTestRunner({
       env,
     })
 
-    apiUtils = axios.create({ baseURL: `http://localhost:${port}` })
+    const cancelTokenSource = axios.CancelToken.source()
+    apiUtils = axios.create({
+      baseURL: `http://localhost:${port}`,
+      cancelToken: cancelTokenSource.token,
+    })
 
     container = container_
-    shutdown = shutdown_
+    shutdown = async () => {
+      await serverShutdown()
+      cancelTokenSource.cancel("Request canceled by shutdown")
+    }
   }
 
   const beforeEach_ = async () => {
+    // The beforeAll already run everything, so lets not re run the loaders for the first iteration
+    if (isFirstTime) {
+      isFirstTime = false
+      return
+    }
+
     const container = options.getContainer()
     const copiedContainer = createMedusaContainer({}, container)