+++ packages/medusa/src/api-v2/admin/inventory-items/middlewares.ts
@@ -1,10 +1,13 @@
 import * as QueryConfig from "./query-config"
 
 import {
+  AdminGetInventoryItemsItemLocationLevelsParams,
   AdminGetInventoryItemsItemParams,
   AdminGetInventoryItemsParams,
   AdminPostInventoryItemsInventoryItemParams,
   AdminPostInventoryItemsInventoryItemReq,
+  AdminPostInventoryItemsItemLocationLevelsLevelParams,
+  AdminPostInventoryItemsItemLocationLevelsLevelReq,
   AdminPostInventoryItemsItemLocationLevelsReq,
   AdminPostInventoryItemsReq,
 } from "./validators"
@@ -41,16 +44,42 @@ export const adminInventoryRoutesMiddlewares: MiddlewareRoute[] = [
   },
   {
     method: ["POST"],
-    matcher: "/admin/inventory-items/:id/location-levels",
-    middlewares: [transformBody(AdminPostInventoryItemsItemLocationLevelsReq)],
-  },
-  {
-    method: ["POST"],
     matcher: "/admin/inventory-items",
     middlewares: [
       transformBody(AdminPostInventoryItemsReq),
       transformQuery(
         AdminGetInventoryItemsItemParams,
+        QueryConfig.retrieveTransformQueryConfig
+      ),
+    ],
+  },
+  {
+    method: ["GET"],
+    matcher: "/admin/inventory-items/:id/location-levels",
+    middlewares: [
+      transformQuery(
+        AdminGetInventoryItemsItemLocationLevelsParams,
+        QueryConfig.listLocationLevelsTransformQueryConfig
+      ),
+    ],
+  },
+  {
+    method: ["POST"],
+    matcher: "/admin/inventory-items/:id/location-levels",
+    middlewares: [transformBody(AdminPostInventoryItemsItemLocationLevelsReq)],
+  },
+  {
+    method: ["POST"],
+    matcher: "/admin/inventory-items/:id/location-levels",
+    middlewares: [transformBody(AdminPostInventoryItemsItemLocationLevelsReq)],
+  },
+  {
+    method: ["POST"],
+    matcher: "/admin/inventory-items/:id/location-levels/:location_id",
+    middlewares: [
+      transformBody(AdminPostInventoryItemsItemLocationLevelsLevelReq),
+      transformQuery(
+        AdminPostInventoryItemsItemLocationLevelsLevelParams,
         QueryConfig.retrieveTransformQueryConfig
       ),
     ],